[WIP] More aggressive clippy
No more clippy warnings.
