+++ src/errors/arithmetic_overflow_error.rs
@@ -1,14 +1,33 @@
-pub mod operands;
+mod operands;
 
 use operands::BinaryOpOperands;
 
-use std::fmt::{Debug, Display, Formatter, Result};
 use std::error::Error;
+use std::fmt::{Debug, Display, Formatter, Result};
+use std::ops::{Add, Mul};
+
+use crate::errors::arithmetic_overflow_error::operands::Operand;
+use crate::errors::arithmetic_overflow_error::operands::add_operands::AddOperands;
+use crate::errors::arithmetic_overflow_error::operands::mul_operands::MulOperands;
 
 #[derive(Debug)]
 pub enum ArithmeticOverflowError {
     Add(Box<dyn BinaryOpOperands>),
     Mul(Box<dyn BinaryOpOperands>),
+}
+
+impl ArithmeticOverflowError {
+    pub fn add_overflow<T: Add + Operand + 'static>(lhs: T, rhs: T) -> Self {
+        Self::Add(Box::new(AddOperands(lhs, rhs)))
+    }
+
+    pub fn mul_overflow<T, F>(lhs: T, rhs: F) -> Self
+    where
+        T: Mul<F> + Operand + 'static,
+        F: Operand + 'static,
+    {
+        Self::Mul(Box::new(MulOperands(lhs, rhs)))
+    }
 }
 
 impl Display for ArithmeticOverflowError {
@@ -19,13 +38,25 @@ impl Display for ArithmeticOverflowError {
     fn fmt(&self, f: &mut Formatter<'_>) -> Result {
         match self {
             Self::Add(operands) => {
-                write!(f, "Adding {} to {} would overflow", operands.lhs(), operands.rhs())
-            },
+                write!(
+                    f,
+                    "Adding {} to {} would overflow",
+                    operands.lhs(),
+                    operands.rhs()
+                )
+            }
             Self::Mul(operands) => {
-                write!(f, "Multiplying {} by {} would overflow", operands.lhs(), operands.rhs())
-            },
+                write!(
+                    f,
+                    "Multiplying {} by {} would overflow",
+                    operands.lhs(),
+                    operands.rhs()
+                )
+            }
         }
     }
 }
 
 impl Error for ArithmeticOverflowError {}
+
+// TODO: Docs, Tests
